requirements.txt
streamlit
endee
groq
pypdf
requests
python-dotenv